Recursive topic function + topic direction
Having Fai asking questions is now a recursive function within the AI
file that stops when it reaches the last question. Fai can also now
detect keyword for each topic and change the topic of conversation
accordingly. Once the topic has been changed it cannot be revisited. At
the moment it only takes one keyword to change the conversation. Further
work needs to be done to revisit topics that were missed if Fai changes
the topic because of keywords.