Missed topic function
Fai will now cycle back to any previous topic that was skipped during
the conversation. I have separated the topic questions into separate
lists to changing topics easier. When going through skipped topics fai
will not change the conversation is keywords are detected. However they
will be collected and stored within the keyword lists. Work that needs
to be done is to find the highest emotional level per topic and then
generate a fortune using that topic and the topics key words that was
collected during the conversation